backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
firebase-admin==6.5.0
google-cloud-pubsub==2.23.0
google-cloud-bigquery==3.25.0
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.0
httpx==0.27.0
pydantic==2.9.0
python-multipart==0.0.9
